+++ scripts/util.js
@@ -1,3 +1,4 @@
+const path = require('path');
 
 const removeUndefinedProperties = (obj) => {
   return Object.keys(obj).reduce((current, key) => {
@@ -9,6 +10,21 @@ const removeUndefinedProperties = (obj) => {
   }, {});
 };
 
+const runParamsConfigModifier = async (params) => {
+  if (!params.configModifier) {
+    return params;
+  }
+  const configModifier = (await import(path.join(process.cwd(), params.configModifier))).default;
+  let newParams = params;
+  if (configModifier.constructor.name === 'AsyncFunction') {
+    newParams = await configModifier(params);
+  } else {
+    newParams = configModifier(params);
+  }
+  return newParams;
+};
+
 module.exports = {
   removeUndefinedProperties,
+  runParamsConfigModifier,
 };